aFull_project/requirements.txt
flask
werkzeug
requests
numpy
scipy
torch
torchvision
opencv-python
Pillow
tqdm
librosa
pickle-mixin
tensorflow
transformers
huggingface-hub
joblib
pandas
google-generativeai
python-dotenv